+++ spec/integration/rest/verify/v2/template.spec.js
@@ -37,6 +37,9 @@ describe('Template', function() {
                   'sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'friendly_name': 'Base Verification Template 2 with do not share',
+                  'channels': [
+                      'sms'
+                  ],
                   'translations': {
                       'en': {
                           'is_default_translation': true,
@@ -71,6 +74,9 @@ describe('Template', function() {
                   'sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'friendly_name': 'Base Verification Template 2 with do not share',
+                  'channels': [
+                      'sms'
+                  ],
                   'translations': {
                       'en': {
                           'is_default_translation': true,
@@ -110,6 +116,9 @@ describe('Template', function() {
                   'sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'friendly_name': 'Base Verification Template 2 with do not share',
+                  'channels': [
+                      'sms'
+                  ],
                   'translations': {
                       'en': {
                           'is_default_translation': true,
@@ -164,6 +173,9 @@ describe('Template', function() {
                   'sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'friendly_name': 'Base Verification Template 2 with do not share',
+                  'channels': [
+                      'sms'
+                  ],
                   'translations': {
                       'en': {
                           'is_default_translation': true,